Add logind inhibit support.
Now, when Inhibit is called, we use a new logind.Conn object to send a
systemd idle inhibit. For now, it's 1:1 - an inhibit to systemd for
every inhibit we get a request for. There is no "crash
detection" (heartbeat) support yet, so we rely on the upstream
inhibitors being well behaved.