Add the scaffolding for heartbeat checking.
Add heartbeatCheck() which will take the mutex every 10 seconds and
iterate over active locks. Right now, it does nothing. Wire up a done
channel such that when Shutdown is called on the bridge we cleanly
stop the goroutine that NewInhibitBridge now starts.

With the peer information stored in lockDetails objects, the next step
will be to ping the peers in the heartbeatCheck() method every time
our ticker fires. This way, we can cleanly drop locks if a peer has
crashed without doing their own cleanup.