Add real heartbeat validation.
When the heartbeat interval is triggered, look up every active peer on
the bus. Then, compare the peer info in our active inhibit locks
against this list. If the peer in the lock isn't in the active set, we
assume they've crashed or shutdown without cleaning up their inhibits
and subsequently clean them up anyway.