Allow toggling verbose logging.
This way we can leave a log file configured but not spam it unless
we're debugging weird interactions with some client that is
misbehaving. There will be a race on the use of *verbose, but that
will never cause any real harm so we don't mutex guard it.

Default logging to false now that we've done this so our default
operation is to not spam the log.