Allow toggling inhibit state from USR1 to allow for keybindings
Instead of toggling logging, which is no longer useful (or potentially
something to do via a systray menu toggle instead), we change USR1
handling behaviour to toggle the "manual inhibit" menu item.

To do this, we add a manualInhibit item to the inhibitBridge object
and store the menu item we create when we initialize the systray. We
add a manualInhibitToggle method to inhibitBridge that sends a message
to the ClickedCh channel of that menu item. By doing this, we reuse
the same code path as a real menu click and let the state management
happen naturally based on the logic there.

There may be a race if the goroutine for the systray hasn't had a
chance to register the manualInhibit menuitem when the first signal is
received, so we do make a nil check that could race. But if the object
is there it should be fine to use the channel. Otherwise, not
signalling it is fine. USR1 signals should be rare and thus even if
this was more risky, it would be quite unlikely to be a problem for
real.

This lets us create a path for adding a sway (or i3 or...) keybinding
to toggle the inhibit state from the keyboard.