Update status after heartbeat checking peers
Using the SDL2 library sees an inhibit placed, but not cleanly removed
unless the programmer knows to make the call that does this. The
bridge would notice the peer had exited and the inhibit would be
cleanly removed, but the status stayed "yellow" (or red if there was
also a manual inhibit) because we didn't make the call to update
status after checking the peers.